requirements.txt
aioschedule~=0.5.2
telebot~=0.0.5
python-dateutil~=2.9.0.post0
pytz~=2024.1
pandas~=2.2.2
requests~=2.31.0
beautifulsoup4~=4.12.3